[fix/#258] 정적 생성 에러 해결
