+++ src/app/forms/created/page.module.scss
@@ -1,7 +1,7 @@
-@use 'src/styles/_variables' as var;
-@use 'src/styles/_color.scss' as color;
-@use 'src/styles/_font' as font;
-@use 'src/styles/_media' as media;
+@use '../../../styles/variables' as var;
+@use '../../../styles/color' as color;
+@use '../../../styles/font' as font;
+@use '../../../styles/media' as media;
 
 .make-your-form {
   position: fixed;
